requirements.txt
anyio==3.6.1
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
colorama==0.4.5
cryptography==38.0.1
dnspython==2.2.1
fastapi==0.85.0
greenlet==1.1.3.post0
h11==0.14.0
httptools==0.5.0
idna==3.4
loguru==0.6.0
motor==3.1.1
passlib==1.7.4
protobuf==3.20.1
pycparser==2.21
pydantic==1.10.2
pymongo==4.3.2
python-dotenv==0.21.0
PyYAML==6.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.20.4
typing_extensions==4.4.0
uvicorn==0.18.3
watchfiles==0.17.0
websockets==10.3
win32-setctime==1.1.0
flask==2.2.2
pyJWT==2.6.0
